Release v0.3.3 - Environment variable auth and documentation updates
Fixes #38 - Environment variable authentication
Fixes #41 - Publisher disconnected status
Fixes #42 - Protocol field mismatch

## Fixed
- Environment variable authentication - Provider now supports:
  - NETSKOPE_SERVER_URL and NETSKOPE_API_KEY (native env vars)
  - TF_VAR_* variables (Terraform standard method)
- Publisher status enum - added 'disconnected' status
- Protocol field mismatch - schema expected 'type' but API returns 'transport'
- Publisher data source path in examples (data[0] → data.publishers[0])
- npa_rules_list response wrapper parsing
- rule_id type mismatch (integer → string)

## Added
- New resource: npa_private_app_public_host
- New data source: npa_publishers_host_os_versions
- Documentation: KNOWN_API_ISSUES.md, OAS_MODIFICATIONS.md
- Post-processing script for Speakeasy regeneration

## Changed
- License changed from Apache 2.0 to BSD 3-Clause
- Updated documentation with both authentication methods

+++ internal/provider/types/protocol_item.go
@@ -8,9 +8,9 @@ import (
 
 type ProtocolItem struct {
 	CreatedAt types.String `tfsdk:"created_at"`
-	ID        types.Int64  `tfsdk:"id"`
+	ID        types.Int32  `tfsdk:"id"`
 	Port      types.String `tfsdk:"port"`
 	Protocol  types.String `tfsdk:"protocol"`
-	ServiceID types.Int64  `tfsdk:"service_id"`
+	ServiceID types.Int32  `tfsdk:"service_id"`
 	UpdatedAt types.String `tfsdk:"updated_at"`
 }
